Added flatbuffer and gtest wraps and added build step for detection list schema.
