+++ subprojects/gtest.wrap
@@ -1,0 +1,16 @@
+[wrap-file]
+directory = googletest-1.15.0
+source_url = https://github.com/google/googletest/archive/refs/tags/v1.15.0.tar.gz
+source_filename = gtest-1.15.0.tar.gz
+source_hash = 7315acb6bf10e99f332c8a43f00d5fbb1ee6ca48c52f6b936991b216c586aaad
+patch_filename = gtest_1.15.0-1_patch.zip
+patch_url = https://wrapdb.mesonbuild.com/v2/gtest_1.15.0-1/get_patch
+patch_hash = 5f8e484c48fdc1029c7fd08807bd2615f8c9d16f90df6d81984f4f292752c925
+source_fallback_url = https://github.com/mesonbuild/wrapdb/releases/download/gtest_1.15.0-1/gtest-1.15.0.tar.gz
+wrapdb_version = 1.15.0-1
+
+[provide]
+gtest = gtest_dep
+gtest_main = gtest_main_dep
+gmock = gmock_dep
+gmock_main = gmock_main_dep
